backend_api/requirements.txt
redis==3.5.3
fastapi==0.68.1
uvicorn==0.15.0
aiohttp==3.7.4
elasticsearch_dsl==7.4.0
elasticsearch==7.10.0
aioredis==2.0.0
fastapi-cache2==0.1.6
orjson==3.6.3
backoff==1.11.1